// src/utils/effect-utils.d.ts
export interface ClipPolygonOptions {
    id: string;
    polygon: GeoJSON.Feature<GeoJSON.Polygon | GeoJSON.MultiPolygon> | GeoJSON.FeatureCollection;
    projection: GeoProjection;
}